refactor(conversations): move panel into features/ behind a shim (Phase 1b)
Relocate the monolithic `Conversations` component to
`app/src/features/conversations/Conversations.tsx`; `pages/Conversations.tsx`
becomes a thin re-export shim so `HumanPage` and the co-located test suites keep
their current import paths during migration (removed in Phase 6). Add
`features/conversations/index.ts` as the public reusability contract.

Resolve the `AgentChatPanel` name collision with the unrelated
`components/settings/panels/AgentChatPanel.tsx`: the page-variant alias is
renamed `ConversationsPage`; `Accounts.tsx` and its webview-selection test mock
updated to match. All relative imports in the moved file deepened one level
(`../` → `../../`); no behavior change.

Typecheck + lint clean; 118 tests (Conversations render/attachments/overflow/
welcomeLock/helpers, HumanPage, Accounts) pass.

Refs: docs/plans/conversations-timeline-refactor.md Phase 1
