+++ app/src/pages/Accounts.tsx
@@ -22,7 +22,7 @@ import {
 } from '../store/mascotSlice';
 import type { Account } from '../types/accounts';
 import { AGENT_ACCOUNT_ID as AGENT_ID } from '../utils/accountsFullscreen';
-import Conversations, { AgentChatPanel } from './Conversations';
+import Conversations, { ConversationsPage } from './Conversations';
 
 // Persistence key for face-toggle state across sessions.
 const FACE_MODE_KEY = 'chat.faceMode';
@@ -172,7 +172,7 @@ const Accounts = () => {
         <main className="relative flex min-w-0 flex-1 flex-col overflow-hidden">
           {isAgentSelected && (
             <div className="min-h-0 flex-1 overflow-hidden">
-              <AgentChatPanel />
+              <ConversationsPage />
             </div>
           )}
         </main>
